On a 64-bit machine, with dev-mode turned on, and pgsql installed, I get warnings that stops the compile. They are fixed now.
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@104073 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ res/res_config_pgsql.c
@@ -921,9 +921,9 @@ static int pgsql_reconnect(const char *database)
 		if (!ast_strlen_zero(dbpass))
 			ast_str_append(&connInfo, 0, " password=%s", dbpass);
 
-		ast_debug(1, "%u connInfo=%s\n", connInfo->len, connInfo->str);
+		ast_debug(1, "%u connInfo=%s\n", (unsigned int)connInfo->len, connInfo->str);
 		pgsqlConn = PQconnectdb(connInfo->str);
-		ast_debug(1, "%u connInfo=%s\n", connInfo->len, connInfo->str);
+		ast_debug(1, "%u connInfo=%s\n", (unsigned int)connInfo->len, connInfo->str);
 		ast_free(connInfo);
 		connInfo = NULL;
